requirements.txt

python-dotenv==1.0.1
pydantic==2.8.2
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==3.0.1
openai==1.40.0
httpx<0.28
pdfminer.six
